Little quick demo of how to use the object position script
Little quick demo of how to use the object position script. It is named in a way that makes it possible to just copy and paste the name of the file into the "from script" field of the script, which makes it super quick to get a nice interactive control null for work in patches.